delete task only user

+++ src/tasks/tasks.service.ts
@@ -28,8 +28,11 @@ export class TasksService {
     return task;
   }
 
-  async deleteTaskById(id: number): Promise<void> {
-    await this.taskRepository.delete(id);
+  async deleteTaskById(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
+    if (result.affected === 0) {
+      throw new NotFoundException(`Task ID ${id} Not found !`);
+    }
   }
 
   async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
